transmission with H.264

+++ src/models/ViTDet/maskedrcnn_vit_b_fpn.py
@@ -247,8 +247,7 @@ class MaskedRCNN_ViT_B_FPN_Contexted(nn.Module):
         new_cache_feature = {}
         
         # convert to tensor
-        # image_tensor = torch.tensor(image_ndarray, dtype=torch.uint8).permute(2, 0, 1).to(self.device)
-        image_tensor = torch.tensor(image_ndarray, dtype=torch.uint8).permute(2, 0, 1).to(self.device).half()
+        image_tensor = torch.tensor(image_ndarray, dtype=torch.uint8).permute(2, 0, 1).to(self.device)
         input = [{"image": image_tensor, "height": image_tensor.shape[-2], "width": image_tensor.shape[-1]}]
         
         # preprocess
